DQ AI Agent/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
great-expectations
langchain
langchain-openai
python-dotenv
